rename files, update references to them
