+++ CONTRIBUTING.md
@@ -25,7 +25,7 @@ tox -p --develop
 ```
 
 ## Meta-tests
-To check that all codes are tested and documented there's a test that error codes mentioned in `README.md`, `CHANGELOG.md` (matching `TRIO\d\d\d`), the keys in `flake8_trio.Error_codes` and codes parsed from filenames and files in `tests/eval_files/`, are all equal.
+To check that all codes are tested and documented there's a test that error codes mentioned in `README.md`, `CHANGELOG.md` (matching `TRIO\d\d\d`), the keys in `flake8_async.Error_codes` and codes parsed from filenames and files in `tests/eval_files/`, are all equal.
 
 ## Test generator
 Tests are automatically generated for files in the `tests/eval_files/` directory, with the code that it's testing interpreted from the file name. The file extension is split off, if there's a match for for `_py\d*` it strips that off and uses it to determine if there's a minimum python version for which the test should only run.
@@ -36,7 +36,7 @@ During tests the result of running the checker on the eval file with autofix ena
 Files without this marker will be checked that they *don't* modify the file content.
 
 ### `error:`
-Lines containing `error:` are parsed as expecting an error of the code matching the file name, with everything on the line after the colon `eval`'d and passed as arguments to `flake8_trio.Error_codes[<error_code>].str_format`. The `globals` argument to `eval` contains a `lineno` variable assigned the current line number, and the `flake8_trio.Statement` namedtuple. The first element after `error:` *must* be an integer containing the column where the error on that line originates.
+Lines containing `error:` are parsed as expecting an error of the code matching the file name, with everything on the line after the colon `eval`'d and passed as arguments to `flake8_async.Error_codes[<error_code>].str_format`. The `globals` argument to `eval` contains a `lineno` variable assigned the current line number, and the `flake8_async.Statement` namedtuple. The first element after `error:` *must* be an integer containing the column where the error on that line originates.
 #### `TRIOxxx:`
 You can instead of `error` specify the error code.
 
@@ -73,6 +73,6 @@ or rote memorization of an arbitrary convention.
 We want to ship bigfixes or new features as soon as they're ready,
 so our release process is automated:
 
-1. Increment `__version__` in `src/flake8_trio.py`
+1. Increment `__version__` in `src/flake8_async.py`
 2. Ensure there's a corresponding entry in `CHANGELOG.md` with same version
 3. Merge to master, and CI will do the rest!
