+++ flake8_async/visitors/visitor111.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import ast
 from typing import TYPE_CHECKING, Any, NamedTuple
 
-from .flake8triovisitor import Flake8AsyncVisitor
+from .flake8asyncvisitor import Flake8AsyncVisitor
 from .helpers import error_class, get_matching_call
 
 if TYPE_CHECKING:
