+++ tox.ini
@@ -24,7 +24,7 @@ commands =
 [pytest]
 addopts =
     --tb=native
-    --cov=flake8_trio
+    --cov=flake8_async
     --cov-branch
     --cov-report=term-missing:skip-covered
     --cov-fail-under=100
@@ -33,8 +33,8 @@ filterwarnings =
 
 [coverage:paths]
 source =
-    flake8_trio
-    */site-packages/flake8_trio
+    flake8_async
+    */site-packages/flake8_async
 
 [coverage:report]
 exclude_lines =
